fix: add stricter compiler warnings and fix type conversion issues
- Add -Wshadow and -Wconversion compiler flags to catch more issues
- Fix implicit int64_t to int conversions in gateway_node.cpp,
  data_access_manager.cpp, and test_gateway_node.cpp using static_cast
- Fix implicit double to float conversions in demo nodes
  (rpm_sensor.cpp, brake_pressure_sensor.cpp)
- Apply clang-format to configuration_manager.cpp, discovery_manager.cpp,
  and rest_server.cpp
- Fix test_integration.test.py to correctly access nested parameter value
  in test_48_delete_configuration_resets_to_default

+++ src/ros2_medkit_gateway/src/data_access_manager.cpp
@@ -32,7 +32,7 @@ DataAccessManager::DataAccessManager(rclcpp::Node * node)
                                                                                                                   "ipt"
                                                                                                                   "s"))
   , native_sampler_(std::make_unique<NativeTopicSampler>(node))
-  , max_parallel_samples_(node->declare_parameter<int>("max_parallel_topic_samples", 10))
+  , max_parallel_samples_(static_cast<int>(node->declare_parameter<int64_t>("max_parallel_topic_samples", 10)))
   , topic_sample_timeout_sec_(node->declare_parameter<double>("topic_sample_timeout_sec", 1.0)) {
   // Validate max_parallel_samples_ against allowed range [1, 50]
   if (max_parallel_samples_ < 1 || max_parallel_samples_ > 50) {
